Refactor OPNet client and routes for improved serialization
Refactored OPNet client to use updated opnet types, improved network handling, and added serialization helpers for Address and contract events. Updated OPNet routes to serialize API responses for receipts, contract data, call results, and public key info. Adjusted transaction parsing to use new opnet transaction types and serialization. Upgraded TypeScript to 5.9.3 in backend dependencies. Minor fixes in pools-updater and file-read utilities.

+++ backend/src/api/bitcoin/bitcoin.routes.ts
@@ -307,7 +307,7 @@ class BitcoinRoutes {
       for (const [index, input] of psbt.data.inputs.entries()) {
         if (!input.nonWitnessUtxo) {
           // Buffer.from ensures it won't be modified in place by reverse()
-          const txid = Buffer.from(psbt.txInputs[index].hash)
+          const txid = Buffer.from(psbt.txInputs[index].hash as Uint8Array)
             .reverse()
             .toString('hex');
